refactor(v2): drop Codacy complexity and long-method warnings
Address the seven Codacy warnings on the v2 design philosophy screens:

  - mock.ts (545 LoC) split into four domain modules under ./mock-data/
    (companies, scenarios, lps, operating); mock.ts becomes a barrel.
  - today.tsx watchList.map (51 LoC) extracted into a typed WatchRow.
  - portfolio.tsx BlotterRow simplified by hoisting className/style
    computations to consts + helper fns; cyclomatic 9 -> ~4.
  - exits.tsx body (110 LoC) decomposed into CinemaMast, CasesGrid,
    CinemaBottom; FUND_RETURNS data-driven instead of duplicated JSX.
  - cash.tsx CashV2 (391 LoC) decomposed into LedgerMast, RibbonSection,
    BreakdownCard, CallsCard, BreakdownRow, CallRow; style constants
    hoisted to module scope.
  - primitives.tsx Sparkline (complexity 10) takes an explicit `endpoint`
    tuple instead of regex-parsing the path; complexity ~2.

No behavior change; visual output is identical. Type-check clean.
